Add C++20/23 compile tests, remove legacy makefiles. Fix warning in MSVC build with latest

+++ SafeInt.hpp
@@ -5804,7 +5804,7 @@ public:
                  std::numeric_limits< T >::is_signed     ? ShiftState_LeftShiftSigned
                                                          : ShiftState_LeftShiftUnsigned,
 
-        isSignedLeftShift = (method == ShiftState_LeftShiftSigned)
+        isSignedLeftShift = (method == (int)ShiftState_LeftShiftSigned)
     };
 };
